+++ src/Client/Credentials/AbstractCredentials.php
@@ -28,6 +28,7 @@ abstract class AbstractCredentials implements CredentialsInterface
         return isset($this->credentials[$offset]);
     }
 
+    #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
         return $this->credentials[$offset];