+++ src/Entity/IterableAPICollection.php
@@ -185,6 +185,7 @@ class IterableAPICollection implements ClientAwareInterface, Iterator, Countable
      * @throws ClientException\Request
      * @throws ClientException\Server
      */
+    #[\ReturnTypeWillChange]
     public function current()
     {
         if (is_null($this->current)) {
@@ -207,6 +208,7 @@ class IterableAPICollection implements ClientAwareInterface, Iterator, Countable
      *
      * @return string
      */
+    #[\ReturnTypeWillChange]
     public function key()
     {
         return