+++ src/Message/InboundMessage.php
@@ -210,6 +210,7 @@ class InboundMessage implements MessageInterface, ArrayAccess, ArrayHydrateInter
      *
      * @throws Exception
      */
+    #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
         trigger_error(